+++ src/PaginationInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Art4\JsonApiClient;
-
-/**
- * Pagination Interface
- */
-interface PaginationInterface extends AccessInterface { }